Rename base class
